KRS Selesai

+++ app/Http/Controllers/Mahasiswa/KrsConstroller.php
@@ -2,11 +2,10 @@
 
 namespace App\Http\Controllers\Mahasiswa;
 
-use App\Models\User;
 use App\Models\Course;
-use App\Models\ProfileUser;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
+use App\Models\SchoolYear;
 use Illuminate\Support\Facades\Auth;
 
 class KrsConstroller extends Controller
@@ -18,13 +17,20 @@ class KrsConstroller extends Controller
      */
     public function index()
     {
-        $program_studi = ProfileUser::where('program_study_id', Auth::user()->profil_user->program_study_id)->first();
-        $courses = Course::where('program_study_id', $program_studi->program_study_id)->latest()->get();
+        // Ambil Data Mata Kuliah Sesuai Program Studi
+        $courses = Course::where('program_study_id', Auth::user()->profil_user->program_study_id)->latest()->get();
 
+        // Cek TA & Semester
+        $school_years = SchoolYear::with('school_year_user')->first();
+        // Cek Setujui
+        $school_year_user = $school_years->school_year_user()->where('user_id', Auth::id())->first();
+
+        // Ambil Data Mata Kuliah User
         $course_users = Course::whereHas('course_user', function ($query) {
-            $query->where('user_id', Auth::id());
+            $query->where('user_id', Auth::id())->where('school_year_id', SchoolYear::first()->id);
         })->get();
-        return view('mahasiswa.krs', compact('courses', 'course_users'));
+
+        return view('mahasiswa.krs', compact('courses', 'course_users', 'school_years', 'school_year_user'));
     }
 
     /**
@@ -35,8 +41,10 @@ class KrsConstroller extends Controller
      */
     public function store(Request $request)
     {
-        $krs = Course::findOrFail($request->course_id);
-        $krs->course_user()->sync(['user_id' => Auth::id()]);
+        $course = Course::with('course_user')->findOrFail($request->course_id);
+        $this->authorize('create', $course);
+
+        $course->course_user()->attach([Auth::id() => ['school_year_id' => SchoolYear::first()->id]]);
         return back()->with('status', 'Data berhasil ditambahkan');
     }
 
@@ -49,7 +57,7 @@ class KrsConstroller extends Controller
     public function destroy($id)
     {
         $krs = Course::findOrFail($id);
-        $krs->course_user()->detach();
+        $krs->course_user()->wherePivot('user_id', Auth::id())->wherePivot('school_year_id', SchoolYear::first()->id)->detach();
         return back()->with('status', 'Data berhasil dihapus');
     }
 }
